ISSUE-370: create a separate class for composer checks [skip ci]
